'CMakeLists.txt' (TrigMenuRulebook-00-04-62)
	* More rules for jet primaries (ATR-13369)
	* TrigMenuRulebook-00-04-62

2016-03-21 Javier Montejo Berlingen
	* Modified rules for jet primaries (ATR-13369)
	* TrigMenuRulebook-00-04-61

2016-03-21 Viviana Cavaliere
	* Fixing a typo in the previous commit
	* TrigMenuRulebook-00-04-60

2016-03-21 Viviana Cavaliere
	* Added 2 new items for the background stream in Standby and Monitoring rules (ATR-13628)
	* Commented for now
	* TrigMenuRulebook-00-04-59

2016-03-18 E. Pianori
	* Enable support items for L1Topo in Cosmic (ATR-13453)
	* TrigMenuRulebook-00-04-58

...
(Long ChangeLog diff - truncated)


Former-commit-id: 78a4bb194f75b8f8d20f33e5d0e121e71fc3295e

+++ Trigger/TriggerCommon/TrigMenuRulebook/python/RuleTools.py
@@ -32,7 +32,7 @@ def calc_eff(d_node, p_node,list_of_streamers):
 
     output = d_node.signature["rate"] *  d_node.signature["prescale"]  / p_node.signature["rate"]
     
-    #Set my hand the efficiency of streamers to 1
+    #Set by hand the efficiency of streamers to 1
     for streamer in list_of_streamers.split(","):        
         if streamer in d_node.name:
             output = 1
@@ -65,12 +65,12 @@ def find_rule(lumi, rule, use_lowest = False):
     else:
        #If the lumi was a string, but isn't matched, then we have a problem
         try:
-            lumi = int(lumi)
+            lumi = float(lumi)
         except ValueError:
             raise ValueError('Non int lumi specified but not found in lumi keys: %s' % lumi)
             #return None
 
-        #Look for the highest rule which is <= lumi and is an int
+        #Look for the highest rule which is <= lumi
         lowest_key = None
         rule_keys = sorted(rule.keys(), reverse=True)
 
@@ -78,7 +78,7 @@ def find_rule(lumi, rule, use_lowest = False):
         for key in rule_keys:
             lowest_key = key
             try:
-                if int(key) <= lumi:
+                if float(key) <= lumi:
                     output_rule = rule[key]
                     output_rule["lumi"] = key
                     assigned_rule = True
